drm/asahi: buffer: Make tvb_something dynamically sized
It looks like this buffer does depend on the tile dimensions, but is
reused from scene to scene. So let's keep it as an Arc in the buffer,
but re-allocate it on new scenes if we need to make it bigger, and have
the scenes also keep a reference to it.

Signed-off-by: Asahi Lina <lina@asahilina.net>