drm/asahi: alloc: Don't leak backing objects
This circular reference thing really needs a rethink. Hopefully it
should go away once VM page tracking is implemented for the new UAPI,
since at that point the VM should hold references to GEM objects, not
the other way around, and we can just let the objects drop when the VM
drops.

Signed-off-by: Asahi Lina <lina@asahilina.net>