soc: apple-pmgr-pwrstate: Add fusing of power states above a level
Read a power state level from 'apple,fuse-states-above' prop in DT, then
logically "fuse" the power states above that level. That is, in changing
of power states to anywhere above the level, accept for the actual state
to be any other power state above the fusing level -- don't distinguish
between them.

In case of MCA power domains, the clocks that the power-domain circuitry
associates with the domains is not the only source of clocking of
the MCA peripheral. As such, the power domain driver will see a clock
gated domain in some modes of MCA operation. Invent this concept of
"fusing of states" to address that.

Signed-off-by: Martin Povišer <povik+lin@cutebit.org>