iommu: apple-dart: Always link consumers with RPM_ACTIVE
Without the RPM_ACTIVE flag on the link, devices lacking runtime PM
support, or devices for which the support is yet to be enabled, are
not considered as active consumers of the DART, and are not preventing
the DART from suspending. Add the new flag to the link and replace an
earlier PCIe-only quirk.

At least one way that DART going through suspend/resume cycle breaks
active consumers is by the DART being reset in the resume path.

Depending on the timing of the iommu probe_device/release_device calls
from within the driver core, this might mean devices without bound
drivers will now keep their DARTs up indefinitely.

Signed-off-by: Martin Povišer <povik+lin@cutebit.org>